+++ Source/SkelGame/Objects/Network/NetworkComponent.h
@@ -30,7 +30,7 @@ namespace Skel
 		GameObject* m_PlayerObjects[Net::MAX_PLAYERS];
 
 		// Server
-		float m_TimeSinceSnapshotSent = 0;
+		double m_TimeSinceSnapshotSent = 0;
 
 		// Client
 		GameObject* m_LocalPlayer;
